refactor
moved the login window to home folder
